fixed graph

+++ app/src/main/java/com/example/stockbroker/Portfolio.java
@@ -2,10 +2,10 @@ package com.example.stockbroker;
 
 public class Portfolio {
     String ticker;
-    Long cost;
+    Double cost;
     Long shares;
 
-    public Portfolio(String ticker, Long shares, Long cost) {
+    public Portfolio(String ticker, Long shares, Double cost) {
         this.ticker = ticker;
         this.shares = shares;
         this.cost = cost;
